// src/enums/status-code.enum.ts
export enum StatusCode {
  STORAGE_NOT_FOUND = 'STORAGE_NOT_FOUND',
  DOWNLOAD_FAILED = 'DOWNLOAD_FAILED',
  PROBE_FAILED = 'PROBE_FAILED',
  NO_VIDEO_TRACK = 'NO_VIDEO_TRACK',
  NO_AUDIO_TRACK = 'NO_AUDIO_TRACK',
  LOW_QUALITY_VIDEO = 'LOW_QUALITY_VIDEO',
  ENCODE_AUDIO_FAILED = 'ENCODE_AUDIO_FAILED',
  ENCODE_VIDEO_FAILED = 'ENCODE_VIDEO_FAILED',
  FINISHED_ENCODING = 'FINISHED_ENCODING',
  CANCELLED_ENCODING = 'CANCELLED_ENCODING'
}